extract FunctionCompiler class
